requirements.txt
Package                            Version
---------------------------------- ------------
matplotlib                         3.7.5
numpy                              1.24.3
pandas                             2.0.3
pillow                             10.3.0
pip                                24.0
PyYAML                             6.0.1
scikit-learn                       1.3.2
scipy                              1.10.1
torch                              1.12.1
tqdm                               4.66.4
